backend/cloud_functions/upload/requirements.txt
functions-framework==3.*
google-cloud-core>=2.3.2
google-cloud-storage>=2.5.0
flask==2.0.2
Werkzeug==2.2.2